feat: interfaces-b09-compat-005 - Prove inventory coverage over baseline aliases
